Refactor groupModal
